Move pixel_to_pixel functionality to astropy.wcs.utils
