+++ astropy/wcs/utils.py
@@ -3,6 +3,7 @@
 import numpy as np
 
 from astropy import units as u
+from astropy.utils.misc import unbroadcast
 
 from .wcs import WCS, WCSSUB_LONGITUDE, WCSSUB_LATITUDE, WCSSUB_CELESTIAL
 
@@ -13,7 +14,7 @@ __all__ = ['add_stokes_axis_to_wcs', 'celestial_frame_to_wcs',
            'proj_plane_pixel_area', 'is_proj_plane_distorted',
            'non_celestial_pixel_scales', 'skycoord_to_pixel',
            'pixel_to_skycoord', 'custom_wcs_to_frame_mappings',
-           'custom_frame_to_wcs_mappings']
+           'custom_frame_to_wcs_mappings', 'pixel_to_pixel']
 
 
 def add_stokes_axis_to_wcs(wcs, add_before_ind):
@@ -632,3 +633,195 @@ def pixel_to_skycoord(xp, yp, wcs, origin=0, mode='all', cls=None):
     coords = cls(frame.realize_frame(data))
 
     return coords
+
+
+def _unique_with_order_preserved(items):
+    """
+    Return a list of unique items in the list provided, preserving the order
+    in which they are found.
+    """
+    new_items = []
+    for item in items:
+        if item not in new_items:
+            new_items.append(item)
+    return new_items
+
+
+def _pixel_to_world_correlation_matrix(wcs):
+    """
+    Return a correlation matrix between the pixel coordinates and the
+    high level world coordinates, along with the list of high level world
+    coordinate classes.
+
+    The shape of the matrix is ``(n_world, n_pix)``, where ``n_world`` is the
+    number of high level world coordinates.
+    """
+
+    # We basically want to collapse the world dimensions together that are
+    # combined into the same high-level objects.
+
+    # Get the following in advance as getting these properties can be expensive
+    all_components = wcs.low_level_wcs.world_axis_object_components
+    all_classes = wcs.low_level_wcs.world_axis_object_classes
+    axis_correlation_matrix = wcs.low_level_wcs.axis_correlation_matrix
+
+    components = _unique_with_order_preserved([c[0] for c in all_components])
+
+    matrix = np.zeros((len(components), wcs.pixel_n_dim), dtype=bool)
+
+    for iworld in range(wcs.world_n_dim):
+        iworld_unique = components.index(all_components[iworld][0])
+        matrix[iworld_unique] |= axis_correlation_matrix[iworld]
+
+    classes = [all_classes[component][0] for component in components]
+
+    return matrix, classes
+
+
+def _pixel_to_pixel_correlation_matrix(wcs_in, wcs_out):
+    """
+    Correlation matrix between the input and output pixel coordinates for a
+    pixel -> world -> pixel transformation specified by two WCS instances.
+
+    The first WCS specified is the one used for the pixel -> world
+    transformation and the second WCS specified is the one used for the world ->
+    pixel transformation. The shape of the matrix is
+    ``(n_pixel_out, n_pixel_in)``.
+    """
+
+    matrix1, classes1 = _pixel_to_world_correlation_matrix(wcs_in)
+    matrix2, classes2 = _pixel_to_world_correlation_matrix(wcs_out)
+
+    if len(classes1) != len(classes2):
+        raise ValueError("The two WCS return a different number of world coordinates")
+
+    # Check if classes match uniquely
+    unique_match = True
+    mapping = []
+    for class1 in classes1:
+        matches = classes2.count(class1)
+        if matches == 0:
+            raise ValueError("The world coordinate types of the two WCS do not match")
+        elif matches > 1:
+            unique_match = False
+            break
+        else:
+            mapping.append(classes2.index(class1))
+
+    if unique_match:
+
+        # Classes are unique, so we need to re-order matrix2 along the world
+        # axis using the mapping we found above.
+        matrix2 = matrix2[mapping]
+
+    elif classes1 != classes2:
+
+        raise ValueError("World coordinate order doesn't match and automatic matching is ambiguous")
+
+    matrix = np.matmul(matrix2.T, matrix1)
+
+    return matrix
+
+
+def _split_matrix(matrix):
+    """
+    Given an axis correlation matrix from a WCS object, return information about
+    the individual WCS that can be split out.
+
+    The output is a list of tuples, where each tuple contains a list of
+    pixel dimensions and a list of world dimensions that can be extracted to
+    form a new WCS. For example, in the case of a spectral cube with the first
+    two world coordinates being the celestial coordinates and the third
+    coordinate being an uncorrelated spectral axis, the matrix would look like::
+
+        array([[ True,  True, False],
+               [ True,  True, False],
+               [False, False,  True]])
+
+    and this function will return ``[([0, 1], [0, 1]), ([2], [2])]``.
+    """
+
+    pixel_used = []
+
+    split_info = []
+
+    for ipix in range(matrix.shape[1]):
+        if ipix in pixel_used:
+            continue
+        pixel_include = np.zeros(matrix.shape[1], dtype=bool)
+        pixel_include[ipix] = True
+        n_pix_prev, n_pix = 0, 1
+        while n_pix > n_pix_prev:
+            world_include = matrix[:, pixel_include].any(axis=1)
+            pixel_include = matrix[world_include, :].any(axis=0)
+            n_pix_prev, n_pix = n_pix, np.sum(pixel_include)
+        pixel_indices = list(np.nonzero(pixel_include)[0])
+        world_indices = list(np.nonzero(world_include)[0])
+        pixel_used.extend(pixel_indices)
+        split_info.append((pixel_indices, world_indices))
+
+    return split_info
+
+
+def pixel_to_pixel(wcs_in, wcs_out, *inputs):
+    """
+    Transform pixel coordinates in a dataset with a WCS to pixel coordinates
+    in another dataset with a different WCS.
+
+    This function is designed to efficiently deal with input pixel arrays that
+    are broadcasted views of smaller arrays, and is compatible with any
+    APE14-compliant WCS.
+
+    Parameters
+    ----------
+    wcs_in : `~astropy.wcs.wcsapi.BaseHighLevelWCS`
+        A WCS object for the original dataset which complies with the
+        high-level shared APE 14 WCS API.
+    wcs_out : `~astropy.wcs.wcsapi.BaseHighLevelWCS`
+        A WCS object for the target dataset which complies with the
+        high-level shared APE 14 WCS API.
+    *inputs :
+        Scalars or arrays giving the pixel coordinates to transform.
+    """
+
+    # Shortcut for scalars
+    if np.isscalar(inputs[0]):
+        world_outputs = wcs_in.pixel_to_world(*inputs)
+        if not isinstance(world_outputs, (tuple, list)):
+            world_outputs = (world_outputs,)
+        return wcs_out.world_to_pixel(*world_outputs)
+
+    # Remember original shape
+    original_shape = inputs[0].shape
+
+    matrix = _pixel_to_pixel_correlation_matrix(wcs_in, wcs_out)
+    split_info = _split_matrix(matrix)
+
+    outputs = [None] * wcs_out.pixel_n_dim
+
+    for (pixel_in_indices, pixel_out_indices) in split_info:
+
+        pixel_inputs = []
+        for ipix in range(wcs_in.pixel_n_dim):
+            if ipix in pixel_in_indices:
+                pixel_inputs.append(unbroadcast(inputs[ipix]))
+            else:
+                pixel_inputs.append(inputs[ipix].flat[0])
+
+        pixel_inputs = np.broadcast_arrays(*pixel_inputs)
+
+        world_outputs = wcs_in.pixel_to_world(*pixel_inputs)
+
+        if not isinstance(world_outputs, (tuple, list)):
+            world_outputs = (world_outputs,)
+
+        pixel_outputs = wcs_out.world_to_pixel(*world_outputs)
+
+        if wcs_out.pixel_n_dim == 1:
+            pixel_outputs = (pixel_outputs,)
+
+        for ipix in range(wcs_out.pixel_n_dim):
+            if ipix in pixel_out_indices:
+                outputs[ipix] = np.broadcast_to(pixel_outputs[ipix], original_shape)
+
+    return outputs[0] if wcs_out.pixel_n_dim == 1 else outputs
